Projetos/ssa-montecarlo-agent/requirements.txt
openai>=1.0.0
gradio
yfinance
numpy
python-dotenv
